+++ src/output.c
@@ -1,0 +1,15 @@
+#include "utils.h"
+#include "stdio.h"
+
+/* append data to output. */
+static void output(mdsys_t *sys, FILE *erg, FILE *traj)
+{
+    int i;
+
+    printf("% 8d % 20.8f % 20.8f % 20.8f % 20.8f\n", sys->nfi, sys->temp, sys->ekin, sys->epot, sys->ekin+sys->epot);
+    fprintf(erg,"% 8d % 20.8f % 20.8f % 20.8f % 20.8f\n", sys->nfi, sys->temp, sys->ekin, sys->epot, sys->ekin+sys->epot);
+    fprintf(traj,"%d\n nfi=%d etot=%20.8f\n", sys->natoms, sys->nfi, sys->ekin+sys->epot);
+    for (i=0; i<sys->natoms; ++i) {
+        fprintf(traj, "Ar  %20.8f %20.8f %20.8f\n", sys->rx[i], sys->ry[i], sys->rz[i]);
+    }
+}
