+++ src/utils.c
@@ -1,0 +1,58 @@
+#include "utils.h"
+#include <sys/time.h>
+#include <stdio.h>
+#include <string.h>
+
+/* helper function: get current time in seconds since epoch */
+static double wallclock()
+{
+        struct timeval t;
+        gettimeofday(&t,0);
+        return ((double) t.tv_sec) + 1.0e-6*((double) t.tv_usec);
+}
+
+/* helper function: zero out an array */
+static void azzero(double *d, const int n)
+{
+    int i;
+    for (i=0; i<n; ++i) {
+        d[i]=0.0;
+    }
+}
+
+/* helper function: apply minimum image convention */
+static double pbc(double x, const double boxby2)
+{
+    while (x >  boxby2) x -= 2.0*boxby2;
+    while (x < -boxby2) x += 2.0*boxby2;
+    return x;
+}
+
+/* helper function: read a line and then return
+   the first string with whitespace stripped off */
+static int get_a_line(FILE *fp, char *buf)
+{
+    char tmp[BLEN], *ptr;
+
+    /* read a line and cut of comments and blanks */
+    if (fgets(tmp,BLEN,fp)) {
+        int i;
+
+        ptr=strchr(tmp,'#');
+        if (ptr) *ptr= '\0';
+        i=strlen(tmp); --i;
+        while(isspace(tmp[i])) {
+            tmp[i]='\0';
+            --i;
+        }
+        ptr=tmp;
+        while(isspace(*ptr)) {++ptr;}
+        i=strlen(ptr);
+        strcpy(buf,tmp);
+        return 0;
+    } else {
+        perror("problem reading input");
+        return -1;
+    }
+    return 0;
+}
